engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclue"
version = "0.1.0"
description = "Terminal-Based Escape Room Engine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0.0",
    "pycryptodome>=3.18.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
audio = ["playsound>=1.3.0"]
dev = ["pyinstaller>=6.0.0", "pytest>=7.0.0"]

[project.scripts]
clue = "clue.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["clue*"]
